Improved a lot support of literals
Literals defined using the Turtle syntax are now recognized properly
which also help determining if a property is an object or a datatype
property.

On the way, the statement lexer (shlex) has been replace by simple splitting
on spaces. Seem ok, but more test would be welcome to ensure minimalkb
behave correctly with malformed statement.

As a side-effect of the new lexer, quoted string are stored as quoted string
and hence kb.lookup can not search anymore in the label. We can certainly
improve that, but I could not quickly find a SQL workaround in sqlite.about()